requirements_dev.txt
numpy 
pandas 
scikit-learn
matplotlib
python-box==6.0.2
pyYAML
ensure
joblib
types-pyYAML
Flask
Flask-Cors
-e .